prevented a character(0) pathname from being insta-read
I added a simple if-statement that wraps the reaction part.
It seemed to be instantly reading a `character(0)` value for `as.character(path$datapath)` because that line is run even before a file can be actually selected. 
Also, I'm not sure why this is all wrapped in an if-statement for `!is.null(input$the_video)` when input literally loads at a value of `0` which is not NULL... Who cares though! Finally works now.